server/requirements.txt
django
djangorestframework
dotenv
mysqlclient

